requirements.txt
flask
jsonschema
Pillow
numpy
ImageHash
opencv-python
Werkzeug